requirements.txt
Flask
python-barcode
reportlab
Pillow
gunicorn
